requirements.txt
amqp==2.3.2
argon2-cffi==18.1.0
billiard==3.5.0.3
celery==4.1.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20121119
Django==2.0.6
django-allauth==0.36.0
django-appconf==1.0.2
django-compressor==2.2
django-crispy-forms==1.7.2
django-debug-toolbar==1.9.1
django-easy-audit==1.0
django-environ==0.4.4
django-extensions==2.0.7
django-extra-views==0.11.0
django-import-export==1.0.1
django-invitations==1.9.2
django-select2==6.0.2
django-shortuuidfield==0.1.3
et-xmlfile==1.0.1
googlemaps==2.5.1
idna==2.6
jdcal==1.4
kombu==4.2.1
oauthlib==2.1.0
odfpy==1.3.6
openpyxl==2.5.4
pycparser==2.18
python3-openid==3.1.0
pytz==2018.4
PyYAML==3.12
rcssmin==1.0.6
requests==2.18.4
requests-oauthlib==1.0.0
rjsmin==1.0.12
shortuuid==0.5.0
six==1.11.0
sqlparse==0.2.4
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.23
vine==1.1.4
xlrd==1.1.0
xlwt==1.3.0
